update ModelSettings.Temperature to type float

+++ SqlSchemaAnalyser/Prompts/PromptConfig.cs
@@ -8,7 +8,7 @@ public record PromptConfig(
 );
 
 public record ModelSettings(
-    double Temperature,
+    float Temperature,
     int MaxTokens
 );
